+++ bundle/src/main/java/com/adobe/acs/commons/redirects/models/HandleQueryString.java
@@ -1,8 +1,9 @@
-/*
- * ACS AEM Commons
- *
- * Copyright (C) 2013 - 2023 Adobe
- *
+/*-
+ * #%L
+ * ACS AEM Commons Bundle
+ * %%
+ * Copyright (C) 2013 - 2025 Adobe
+ * %%
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -14,6 +15,26 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ * #L%
  */
-@org.osgi.annotation.versioning.Version("6.0.4")
-package com.adobe.acs.commons.redirects.filter;
+package com.adobe.acs.commons.redirects.models;
+
+/**
+ * how to handle query strings in the redirect
+ */
+public enum HandleQueryString {
+  /**
+   * ignore, i.e. don't append query string from the request
+   */
+  IGNORE,
+
+  /**
+   * Replace query string in the target with query string in the request
+   */
+  REPLACE,
+
+  /**
+   * Combine query string in the target with query string in the request
+   */
+  COMBINE
+}