+++ bundle/src/main/java/com/adobe/acs/commons/redirects/servlets/ExportRedirectMapServlet.java
@@ -118,6 +118,7 @@ public class ExportRedirectMapServlet extends SlingSafeMethodsServlet {
         headerRow.createCell(11).setCellValue(ExportColumn.MODIFIED.getTitle());
         headerRow.createCell(12).setCellValue(ExportColumn.MODIFIED_BY.getTitle());
         headerRow.createCell(13).setCellValue(ExportColumn.CACHE_CONTROL.getTitle());
+        headerRow.createCell(14).setCellValue(ExportColumn.PRESERVE_QUERY_STRING.getTitle());
 
         // column width in POI is measured in 1/256th of the default character width
         sheet.setColumnWidth(0, 256 * 50);
@@ -134,6 +135,7 @@ public class ExportRedirectMapServlet extends SlingSafeMethodsServlet {
         sheet.setColumnWidth(11, 256 * 12);
         sheet.setColumnWidth(12, 256 * 30);
         sheet.setColumnWidth(13, 256 * 30);
+        sheet.setColumnWidth(14, 256 * 12);
 
         for (Cell cell : headerRow) {
             cell.setCellStyle(headerStyle);
@@ -186,6 +188,9 @@ public class ExportRedirectMapServlet extends SlingSafeMethodsServlet {
 
             Cell cell11 = row.createCell(13);
             cell11.setCellValue(rule.getCacheControlHeader());
+
+            Cell cell14 = row.createCell(14);
+            cell14.setCellValue(rule.getPreserveQueryString());
         }
         sheet.setAutoFilter(new CellRangeAddress(0, rownum - 1, 0, 13));