+++ CHANGELOG.md
@@ -6,7 +6,7 @@ in https://github.com/Adobe-Consulting-Services/acs-aem-commons/releases.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com)
 
-<!-- Keep this up to date! After a release, change the tag name to the latest release -->- 
+<!-- Keep this up to date! After a release, change the tag name to the latest release -->-
 
 ## Unreleased ([details][unreleased changes details])
  
@@ -36,6 +36,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 ### Changed
 
 - #3494 - Remove offline instrumentation with Jacoco
+- #3509 - Redirect Manager: support combining query string in the target with query string in the request
 
 ### Fixed